+++ ng-directive/templates/new-directive/_new-page.spec.js
@@ -1,24 +1,27 @@
 'use strict';
 
 describe('<%= optModulePrefix %><%= optDirectiveNameCamel %>', function() {
-	var elm, scope, $compile, $timeout;
+	var elm, elmScope, $scope, $compile, $timeout;
 	
-	// beforeEach(module('myApp'));
-	beforeEach(module('app'));
+	beforeEach(module('myApp'));
 	
 	/**
 	@param params
+		@param {String} html
 	*/
 	var createElm =function(params) {
-		var html ="<div><div <%= optModulePrefix %>-<%= optDirectiveName %>>"+
-		"</div></div>";
-		elm =angular.element(html);
-		
-		$compile(elm)(scope);
-		scope.$digest();
-		scope.$apply();		//NOTE: required otherwise the directive won't be compiled!!!! ... wtf?
+		var html ="<div <%= optModulePrefix %>-<%= optDirectiveName %>>"+
+		"</div>";
+		if(params.html) {
+			html =params.html;
+		}
+		// elm =angular.element(html);
+		elm =$compile(html)($scope);
+		// $scope.$digest();
+		$scope.$apply();		//NOTE: required otherwise the alert directive won't be compiled!!!! ... wtf?
+		elmScope =elm.isolateScope();
 		var elements ={
-			'<%= optDirectiveName %>':elm.find('div').children().find('div')
+			// 'somePart':elm.find('div').children().find('div')
 		};
 		return elements;
 	};
@@ -26,11 +29,11 @@ describe('<%= optModulePrefix %><%= optDirectiveNameCamel %>', function() {
 	beforeEach(inject(function(_$rootScope_, _$compile_, _$timeout_) {
 		$compile = _$compile_;
 		$timeout = _$timeout_;
-		scope = _$rootScope_.$new();
+		$scope = _$rootScope_.$new();
 	}));
 	
-	afterEach(function() {
-	});
+	// afterEach(function() {
+	// });
 	
 	/*
 	it('should do something', function() {