+++ ng-route/templates/new-page/_NewPageCtrl.spec.js
@@ -1,13 +1,13 @@
 'use strict';
 
-describe('<%= optRouteNameCtrl %>', function(){
-	var ctrl, scope ={};
+describe('<%= optRouteNameCtrl %>', function() {
+	var $ctrl, $scope ={};
 	
 	beforeEach(module('myApp'));
 	
-	beforeEach(inject(function($rootScope, $controller) {
-		scope = $rootScope.$new();
-		ctrl = $controller('<%= optRouteNameCtrl %>', {$scope: scope});
+	beforeEach(inject(function(_$rootScope_, _$controller_) {
+		$scope = _$rootScope_.$new();
+		$ctrl = _$controller_('<%= optRouteNameCtrl %>', {$scope: $scope});
 	}));
 	
 	/*